mavlink_main: fix CAMERA_TRIGGER stream
The CAMERA_TRIGGER stream uses a constant stream size, and thus needs to
use the default stream rate, which is -1 (meaning as fast as possible, or
whenever the topic is updated in that case).

By using a rate of of 500 Hz, the rate computation broke: mavlink thought
there was a constant bandwidth requirement of 500 * camera trigger message
size.