sensors/ROMFS: Use adc for battery voltage in HITL
Instead of just hardcoding 11.5 volts, it's nicer to have the actual
voltage in hardware in the loop. This way low-voltage actions can be
simulated with a power module. Also, it means that the H520 will shut
off when the battery is empty which is a nice to have.

Being able to programatically set the voltage to simulate low voltage
would also be nice but that's omitted for now.