commander: Removed yuneec factory calibration mode
the factory calibration mode was added here:
https://github.com/YUNEEC/Firmware/commit/b3fc2b676726fa699ee8fffcc19286a53999d326

It is no longer compatible since upstream removed the vehicle_attitude
subscription from the commander (rightfully so) and this block should be moved
somewhere else if it still needed. Otherwise we have to add the attitude
subscription back to the commander, but that's a bad idea. We have to keep
upstream diffs minimal.