Battery: switch to dynamic remaining time weight
Because there's no easy way to initialize the estimated flight time
without smart battery and there's a filter acting on it, there's a
convergence phanse in the beginning of each flight. To accelerate
the convergence I made the weight of the filter dependent on the
difference between the current time estimate and the filtered one.