st24: add check for M4 raw channel mapping version
In channel 9 the mapping version is embedded. I add a check for it
and if the version is not compatible I produce an RC loss.
I still need to add a message to the user otherwise the error will
not be found and solved by updating the ST16 to the correct version.