mc_pos_control: add turtle slider to scale horizontal velocity
With the new raw mapping the hardcoded turtle slider scaling is
finally gone and we can decide directly during consumation of the sticks
what to do with the slider data. For now we control maximal velocity
in only position control mode to enable expected legacy functionality.

Tilt of manual mode does not get limited more with the slider anymore.