yuneec_tunes: make low battery alarm non-repeating
Before the repeating tunes were broken and as a result there was no
beeping with low battery. Now it got fixed and we have the super
annoying never ending beep with low battery which is not desirable
for user experience. This is a quick fix that should output the same
default beeps but only for a limited time.