mc_pos_control: workaround for landing gear
The landing gear was only handeled acceptable because of a huge hack in
the position controller. This commit puts back a workaround to have it
working again. There needs to be a better solution soon but we need
legacy functionality for the releases and switching to flight tasks will
simplify the problem resolution a lot.