ROMFS: customize telemetry rates to improve Exif
Previously, the global GLOBAL_POSITION_INT topic was only sent to the
gimbal and camera with 5 Hz, leading to an inaccuracy of up to
2 meters at a velocity of 10 m/s.
This means that 3D reconstruction (e.g. using Pix4D) would result in
large reprojection errors.

By reducing some of the (unused) telemetry fields and increasing the
rate of GLOBAL_POSITION_INT to 50 Hz, we should be able to reduce the
error to 0.2 meters.