tap-v3: use board define instead of hwlabelcmp
It is easiest to just use a define for the ofdm board and then use
that one define everywhere. This way we don't need to have state in fmu
and in the ver command but can just rely on how the fmu driver was
built.

This removes board specific things from makefiles and moves it all to
the actual board specific cmake file.