turtle_slider scaling stick input directly instead of velocity.
If we scale the final velocity like before this commit, we will scale again after
the limits of optical-flow have already been applied. Meaning, that flow could for
example reduce the velocity from 10 m/2 to 4 m/2, and then the turtle slider
reduces it to 10%, which is 0.4 m/s when in fact it should be 10% of the originally
defined velocity.