vmount: try to fix absolute gimbal angle
From testing it looks like that the absolute gimbal angle depends on the
orientation of the drone when it is switched on.

The assumption for this fix is that the gimbal initializes its yaw angle
incorrectly because during/before EKF initialization of the autopilot a
wrong attitude is sent to the gimbal via the gimbal_control message.
Bench testing showed that a quaterion of (0, 0, 0, 0) is sent before the
EKF has aligned. If the gimbal interprets this garbage angle and sets
the current angle e.g. as North that would explain why the absolute
angle is off later.

This issue would only happen if the ST10C is switched on before the EKF
is aligned because gimbal_control messages are only sent once RC is
connected.

Edit: after testing it turned out that the gimbal firmware does not
actually support an absolute angle, so this change does not fix it.
However, it probably makes sense to keep it because it avoids sending
garbage before initialization.