Removed failure check for diagonal motor
When a motor failure is detected, always run FTC.
Previously the driver would check if the diagonally
opposed motor was fault-free and only then FTC would run.
However, I believe there is nothing to lose if we try to
control the vehicle with 5 rotors. Even though if more
than one motor show errors. It's the best we can do.