Don't restart failing motors.
Previously the driver would restart failing motors if
they fail while FTC is already active. There was logic
in the driver, that would assume that a lose propeller
can touch the propellers of neighbouring motors and it
would send RPMSTOPPED to those motors and then restart
them. I removed all of it. There is now only one motor
that can be stopped because of FTC, and that's the
first failure! Everything else is ignored.