mc_att_control: hotfix to prevent yaw wobbling with big steps
This can only be seen in large yaw steps in auto. I tested the
controller itself even with high yaw weight does not cause this wobbling
so there has to be something wrong with the setpoint generation.
I suspect a dependency of vehicle attitude (also roll, pitch) in the
setpoint generation and hence an additional loop which is oscillating.
Needs more time to debug.